// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Message = {
    ok: boolean;
    message: string;
};

export type TokenRequest = {
    password: string;
};

export type TokenResponse = {
    token: string;
};

export type Notes = {
    notes: Array<(string)>;
    nextPage?: string;
};

export type Note = {
    key: string;
    body: string;
    modified?: string;
};

export type GetHealthResponse = Message;

export type PostTokenData = {
    requestBody?: TokenRequest;
};

export type PostTokenResponse = TokenResponse | Message;

export type GetNoteData = {
    authorization: string;
    dir?: string;
    page?: string;
    q?: string;
};

export type GetNoteResponse = Notes | Message;

export type PostNoteData = {
    authorization: string;
    requestBody?: Note;
};

export type PostNoteResponse = Note | Message;

export type GetNoteByKeyData = {
    authorization: string;
    key: string;
};

export type GetNoteByKeyResponse = Note | Message;

export type PutNoteByKeyData = {
    authorization: string;
    key: string;
    requestBody?: Note;
};

export type PutNoteByKeyResponse = Note | Message;

export type $OpenApiTs = {
    '/health': {
        get: {
            res: {
                default: Message;
            };
        };
    };
    '/token': {
        post: {
            req: PostTokenData;
            res: {
                200: TokenResponse;
                default: Message;
            };
        };
    };
    '/note': {
        get: {
            req: GetNoteData;
            res: {
                200: Notes;
                default: Message;
            };
        };
        post: {
            req: PostNoteData;
            res: {
                200: Note;
                default: Message;
            };
        };
    };
    '/note/{key}': {
        get: {
            req: GetNoteByKeyData;
            res: {
                200: Note;
                default: Message;
            };
        };
        put: {
            req: PutNoteByKeyData;
            res: {
                200: Note;
                default: Message;
            };
        };
    };
};